provide a textarea for pasting a custom file, hidden in a detail element
For that, move the LocalStorage fs into the MountableFileSystem, so it can exist in parallel to the ZipFS one. Move the samples to /samples and add a shortcut command to list those files.

Modify the command-click handler to match all clicks with data-command attributes, not just inside #examples. Fixes the bug where clicking around the sample buttons would run an empty-string command.

Since there's now more commands, optimize some labels and the layout, also making it wrap on smaller screens