implement a custom hook to handle panics
This hook is very similar to the console_error_panic_hook, but it instead binds to a JS module that the "app" exposes and which can show the panic message to the user, along with a shortcut to recover the broken process.

This is still crude, but a huge improvement over the complete lack of feedback (when not looking at devtools).